int na double

+++ main.cpp
@@ -1,7 +1,7 @@
 #include <iostream>
 using namespace std;
 
-int main()
+double main()
 {
     cout << "Enter A and B: ";
     int a, b;
